+++ include/pybind/adapter/sequence_get_adapter.hpp
@@ -24,7 +24,7 @@ namespace pybind
         }
 
     public:
-        virtual PyObject * call( kernel_interface * _kernel, void * _self, const class_type_scope_interface_ptr & scope, uint32_t _index ) = 0;
+        virtual PyObject * call( kernel_interface * _kernel, void * _self, const class_type_scope_interface_ptr & scope, size_t _index ) = 0;
     };
     //////////////////////////////////////////////////////////////////////////
     typedef intrusive_ptr<sequence_get_adapter_interface> sequence_get_adapter_interface_ptr;
@@ -45,7 +45,7 @@ namespace pybind
         }
 
     protected:
-        PyObject * call( kernel_interface * _kernel, void * _impl, const class_type_scope_interface_ptr & _scope, uint32_t _index ) override
+        PyObject * call( kernel_interface * _kernel, void * _impl, const class_type_scope_interface_ptr & _scope, size_t _index ) override
         {
             C * self = _kernel->meta_cast_class_t<C>( _impl, _scope );
 
@@ -76,7 +76,7 @@ namespace pybind
         }
 
     protected:
-        PyObject * call( kernel_interface * _kernel, void * _impl, const class_type_scope_interface_ptr & _scope, uint32_t _index ) override
+        PyObject * call( kernel_interface * _kernel, void * _impl, const class_type_scope_interface_ptr & _scope, size_t _index ) override
         {
             const char * scopeName = _kernel->get_class_type_info_t<C>();
 
@@ -116,7 +116,7 @@ namespace pybind
         }
 
     protected:
-        PyObject * call( kernel_interface * _kernel, void * _impl, const class_type_scope_interface_ptr & _scope, uint32_t _index ) override
+        PyObject * call( kernel_interface * _kernel, void * _impl, const class_type_scope_interface_ptr & _scope, size_t _index ) override
         {
             C * self = _kernel->meta_cast_class_t<C>( _impl, _scope );
 
@@ -148,7 +148,7 @@ namespace pybind
         }
 
     protected:
-        PyObject * call( kernel_interface * _kernel, void * _impl, const class_type_scope_interface_ptr & _scope, uint32_t _index ) override
+        PyObject * call( kernel_interface * _kernel, void * _impl, const class_type_scope_interface_ptr & _scope, size_t _index ) override
         {
             C * self = _kernel->meta_cast_class_t<C>( _impl, _scope );
 
@@ -178,7 +178,7 @@ namespace pybind
         }
 
     protected:
-        PyObject * call( kernel_interface * _kernel, void * _impl, const class_type_scope_interface_ptr & _scope, uint32_t _index ) override
+        PyObject * call( kernel_interface * _kernel, void * _impl, const class_type_scope_interface_ptr & _scope, size_t _index ) override
         {
             C * self = _kernel->meta_cast_class_t<C>( _impl, _scope );
